test-requirements.txt
#Libraries for testing
sure
selenium == 2.42.1
flask-testing